add same logic also for a plugin install
