+++ setup/create-wp-project/src/commands/theme.js
@@ -27,6 +27,7 @@ const { scriptArguments } = require('../arguments');
 const { installModifiedNodeDependencies, installModifiedComposerDependencies } = require('../dependencies');
 const { alertBox } = require('../basics/misc');
 const chalk = require('chalk');
+const { navigateToDirectory } = require('../navigate-to-directory');
 
 exports.command = ['*', 'theme'];
 exports.desc = 'Setup a new WordPress theme. Should be run inside your theme folder (wp-content/themes).';
@@ -36,38 +37,8 @@ exports.handler = async (argv) => {
 	await clearConsole();
 	await writeIntro();
 
-	const fs = require('fs');
-
-	const findProjectRoot = (currentDir) => {
-		if (!currentDir || currentDir === '/') {
-			// Reached the root without finding wp-config.php
-			return null;
-		}
-
-		const wpConfigPath = path.join(currentDir, 'wp-config.php');
-		const wpConfigExists = fs.existsSync(wpConfigPath);
-
-		if (wpConfigExists) {
-			// Found wp-config.php, return the current directory as the project root
-			return currentDir;
-		}
-
-		// Move up one directory level and continue the search
-		const parentDir = path.dirname(currentDir);
-		return findProjectRoot(parentDir);
-	};
-
-	// Find the root directory of the WordPress project
-	const rootDir = findProjectRoot(process.cwd());
-
-	if (!rootDir) {
-		console.error('Unable to find project root.');
-		process.exit(1);
-	}
-
-	// The root directory, navigate to the themes folder
-	const themesDir = path.join(rootDir, 'wp-content', 'themes');
-	process.chdir(themesDir);
+	// Navigate to the themes directory
+	navigateToDirectory('themes');
 
 	// Trigger prompts if any of the required arguments was not set
 	const promptedInfo = await maybePrompt(scriptArguments, argv);
